Separated generators into install and cable
