+++ lib/generators/install/templates/application_cable_channel.rb.erb
