+++ lib/generators/install/templates/application_cable_connection.rb.erb
