+++ lib/generators/install/templates/cable.yml.erb
@@ -1,0 +1,14 @@
+# production: &production
+#   :url: redis://localhost:6379
+#   :host: localhost
+#   :port: 6379
+#   :timeout: 1
+
+local: &local
+  :url: redis://localhost:6379
+  :host: localhost
+  :port: 6379
+  :timeout: 1
+  :inline: true
+development: *local
+test: *local
