+++ lib/generators/install/templates/index.coffee.erb
